backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
cryptography>=42.0.0
httpx>=0.26.0
facebook-business>=19.0.0
celery[redis]>=5.3.0
redis>=5.0.0
structlog>=24.1.0
numpy>=1.26.0
uuid-utils>=0.6.1
python-multipart>=0.0.6
weasyprint>=59.0
jinja2>=3.1.0
gspread>=5.10.0
google-auth>=2.0.0

# Dev / test
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
testcontainers[postgres,redis]>=3.7.0
pyyaml>=6.0.0
